requirements.txt
opencv-python
requests
pillow
python-dotenv 